organizing the code
